Automatically Set Channel Bitrate
